Trivial stemming algorithm for geman language
Added a trivial stemming algorithm to include flections of words (such as [des] "Bundespräsidenten" when searching for "Bundespräsident"). The method is not guaranteed to result in a meaningful word ([des] Norwegers becomes "Norweg"), but the real objective is that (ideally) all flections become _the same_ (albeit non-existent) word.